+++ SofaKernel/modules/SofaBaseCollision/DefaultPipeline.cpp
@@ -36,8 +36,9 @@
 #endif
 
 #include <sofa/helper/AdvancedTimer.h>
+using sofa::helper::AdvancedTimer ;
+using sofa::helper::ScopedAdvancedTimer ;
 
-#define VERBOSE(a) if (bVerbose.getValue()) a; else {}
 
 namespace sofa
 {
@@ -61,9 +62,13 @@ int DefaultPipelineClass = core::RegisterObject("The default collision detection
         ;
 
 DefaultPipeline::DefaultPipeline()
-    : bVerbose(initData(&bVerbose, false, "verbose","Display current step information"))
-    , bDraw(initData(&bDraw, false, "draw","Draw detected collisions"))
-    , depth(initData(&depth, 6, "depth","Max depth of bounding trees"))
+    : d_doPrintInfoMessage(initData(&d_doPrintInfoMessage, false, "verbose",
+                                    "Display extra informations at each computation step. (default=false)"))
+    , d_doDebugDraw(initData(&d_doDebugDraw, false, "draw",
+                             "Draw the detected collisions. (default=false)"))
+    //TODO(dmarchal) fix the min & max value with response from discussion.
+    , d_depth(initData(&d_depth, 6, "depth",
+                       "Max depth of bounding trees. (default=6, min=?, max=?)"))
 {
 }
 
@@ -73,50 +78,58 @@ typedef simulation::Visitor::ctime_t ctime_t;
 
 void DefaultPipeline::doCollisionReset()
 {
-    VERBOSE(sout << "DefaultPipeline::doCollisionReset, Reset collisions"<<sendl);
+    msg_info_when(d_doPrintInfoMessage.getValue())
+            << "DefaultPipeline::doCollisionReset" ;
+
     core::objectmodel::BaseContext* scene = getContext();
     // clear all contacts
-    if (contactManager!=NULL)
+    if (contactManager!=nullptr)
     {
-        const sofa::helper::vector<Contact::SPtr>& contacts = contactManager->getContacts();
-        for (sofa::helper::vector<Contact::SPtr>::const_iterator it = contacts.begin(); it!=contacts.end(); ++it)
+        const helper::vector<Contact::SPtr>& contacts = contactManager->getContacts();
+        for (helper::vector<Contact::SPtr>::const_iterator it = contacts.begin(); it!=contacts.end(); ++it)
         {
             (*it)->removeResponse();
         }
     }
     // clear all collision groups
-    if (groupManager!=NULL)
+    if (groupManager!=nullptr)
     {
         groupManager->clearGroups(scene);
     }
 }
 
-void DefaultPipeline::doCollisionDetection(const sofa::helper::vector<core::CollisionModel*>& collisionModels)
+void DefaultPipeline::doCollisionDetection(const helper::vector<core::CollisionModel*>& collisionModels)
 {
-    sofa::helper::AdvancedTimer::stepBegin("doCollisionDetection");
-    VERBOSE(sout << "DefaultPipeline::doCollisionDetection, Compute Bounding Trees"<<sendl);
+    ScopedAdvancedTimer docollisiontimer("doCollisionDetection");
+
+    msg_info_when(d_doPrintInfoMessage.getValue())
+         << "doCollisionDetection, compute Bounding Trees" ;
+
     // First, we compute a bounding volume for the collision model (for example bounding sphere)
     // or we have loaded a collision model that knows its other model
 
-    sofa::helper::vector<CollisionModel*> vectBoundingVolume;
+    helper::vector<CollisionModel*> vectBoundingVolume;
     {
-        sofa::helper::AdvancedTimer::stepBegin("BBox");
+        ScopedAdvancedTimer bboxtimer("BBox");
+
 #ifdef SOFA_DUMP_VISITOR_INFO
         simulation::Visitor::printNode("ComputeBoundingTree");
 #endif
         const bool continuous = intersectionMethod->useContinuous();
         const SReal dt       = getContext()->getDt();
 
-        sofa::helper::vector<CollisionModel*>::const_iterator it;
-        const sofa::helper::vector<CollisionModel*>::const_iterator itEnd = collisionModels.end();
+        helper::vector<CollisionModel*>::const_iterator it;
+        const helper::vector<CollisionModel*>::const_iterator itEnd = collisionModels.end();
         int nActive = 0;
 
         for (it = collisionModels.begin(); it != itEnd; ++it)
         {
-            VERBOSE(sout << "DefaultPipeline::doCollisionDetection, consider model "<<(*it)->getName()<<sendl);
+            msg_info_when(d_doPrintInfoMessage.getValue())
+                << "doCollisionDetection, consider model" ;
+
             if (!(*it)->isActive()) continue;
 
-            int used_depth = broadPhaseDetection->needsDeepBoundingTree() ? depth.getValue() : 0;
+            int used_depth = broadPhaseDetection->needsDeepBoundingTree() ? d_depth.getValue() : 0;
 
             if (continuous)
                 (*it)->computeContinuousBoundingTree(dt, used_depth);
@@ -126,71 +139,82 @@ void DefaultPipeline::doCollisionDetection(const sofa::helper::vector<core::Coll
                 vectBoundingVolume.push_back ((*it)->getFirst());
             ++nActive;
         }
-        sofa::helper::AdvancedTimer::stepEnd("BBox");
+
 #ifdef SOFA_DUMP_VISITOR_INFO
         simulation::Visitor::printCloseNode("ComputeBoundingTree");
 #endif
-        VERBOSE(sout << "DefaultPipeline::doCollisionDetection, Computed "<<nActive<<" BBoxs"<<sendl);
+
+        msg_info_when(d_doPrintInfoMessage.getValue())
+                << "doCollisionDetection, Computed "<<nActive<<" BBoxs" ;
     }
     // then we start the broad phase
-    if (broadPhaseDetection==NULL) return; // can't go further
-    VERBOSE(sout << "DefaultPipeline::doCollisionDetection, BroadPhaseDetection "<<broadPhaseDetection->getName()<<sendl);
+    if (broadPhaseDetection==nullptr) return; // can't go further
+
+    msg_info_when(d_doPrintInfoMessage.getValue())
+            << "doCollisionDetection, BroadPhaseDetection "<<broadPhaseDetection->getName();
+
 #ifdef SOFA_DUMP_VISITOR_INFO
     simulation::Visitor::printNode("BroadPhase");
 #endif
-    sofa::helper::AdvancedTimer::stepBegin("BroadPhase");
-    intersectionMethod->beginBroadPhase();
-    broadPhaseDetection->beginBroadPhase();
-    broadPhaseDetection->addCollisionModels(vectBoundingVolume);  // detection is done there
-    broadPhaseDetection->endBroadPhase();
-    intersectionMethod->endBroadPhase();
-    sofa::helper::AdvancedTimer::stepEnd  ("BroadPhase");
-
+    {
+        ScopedAdvancedTimer broadphase("BroadPhase");
+        intersectionMethod->beginBroadPhase();
+        broadPhaseDetection->beginBroadPhase();
+        broadPhaseDetection->addCollisionModels(vectBoundingVolume);  // detection is done there
+        broadPhaseDetection->endBroadPhase();
+        intersectionMethod->endBroadPhase();
+    }
 #ifdef SOFA_DUMP_VISITOR_INFO
     simulation::Visitor::printCloseNode("BroadPhase");
 #endif
 
     // then we start the narrow phase
-    if (narrowPhaseDetection==NULL) return; // can't go further
-    VERBOSE(sout << "DefaultPipeline::doCollisionDetection, NarrowPhaseDetection "<<narrowPhaseDetection->getName()<<sendl);
+    if (narrowPhaseDetection==nullptr) return; // can't go further
+
+    msg_info_when(d_doPrintInfoMessage.getValue())
+        << "doCollisionDetection, NarrowPhaseDetection "<<narrowPhaseDetection->getName();
 
 #ifdef SOFA_DUMP_VISITOR_INFO
     simulation::Visitor::printNode("NarrowPhase");
 #endif
-    sofa::helper::AdvancedTimer::stepBegin("NarrowPhase");
-    intersectionMethod->beginNarrowPhase();
-    narrowPhaseDetection->beginNarrowPhase();
-    sofa::helper::vector<std::pair<CollisionModel*, CollisionModel*> >& vectCMPair = broadPhaseDetection->getCollisionModelPairs();
-    VERBOSE(sout << "DefaultPipeline::doCollisionDetection, "<< vectCMPair.size()<<" colliding model pairs"<<sendl);
-    narrowPhaseDetection->addCollisionPairs(vectCMPair);
-    narrowPhaseDetection->endNarrowPhase();
-    intersectionMethod->endNarrowPhase();
-    sofa::helper::AdvancedTimer::stepEnd  ("NarrowPhase");
+    {
+        ScopedAdvancedTimer narrowphase("NarrowPhase");
+        intersectionMethod->beginNarrowPhase();
+        narrowPhaseDetection->beginNarrowPhase();
+        helper::vector<std::pair<CollisionModel*, CollisionModel*> >& vectCMPair = broadPhaseDetection->getCollisionModelPairs();
 
+        msg_info_when(d_doPrintInfoMessage.getValue())
+                << "doCollisionDetection, "<< vectCMPair.size()<<" colliding model pairs" ;
+
+        narrowPhaseDetection->addCollisionPairs(vectCMPair);
+        narrowPhaseDetection->endNarrowPhase();
+        intersectionMethod->endNarrowPhase();
+    }
 #ifdef SOFA_DUMP_VISITOR_INFO
     simulation::Visitor::printCloseNode("NarrowPhase");
 #endif
 
-    sofa::helper::AdvancedTimer::stepEnd("doCollisionDetection");
 }
 
 void DefaultPipeline::doCollisionResponse()
 {
     core::objectmodel::BaseContext* scene = getContext();
     // then we start the creation of contacts
-    if (contactManager==NULL) return; // can't go further
-    VERBOSE(sout << "Create Contacts "<<contactManager->getName()<<sendl);
+    if (contactManager==nullptr) return; // can't go further
+
+    msg_info_when(d_doPrintInfoMessage.getValue())
+        << "Create Contacts "<<contactManager->getName() ;
+
     contactManager->createContacts(narrowPhaseDetection->getDetectionOutputs());
 
     // finally we start the creation of collisionGroup
 
-    const sofa::helper::vector<Contact::SPtr>& contacts = contactManager->getContacts();
+    const helper::vector<Contact::SPtr>& contacts = contactManager->getContacts();
 
     // First we remove all contacts with non-simulated objects and directly add them
-    sofa::helper::vector<Contact::SPtr> notStaticContacts;
+    helper::vector<Contact::SPtr> notStaticContacts;
 
-
-    for (sofa::helper::vector<Contact::SPtr>::const_iterator it = contacts.begin(); it!=contacts.end(); ++it)
+    for (helper::vector<Contact::SPtr>::const_iterator it = contacts.begin(); it!=contacts.end(); ++it)
     {
         Contact::SPtr c = *it;
         if (!c->getCollisionModels().first->isSimulated())
@@ -205,18 +229,21 @@ void DefaultPipeline::doCollisionResponse()
             notStaticContacts.push_back(c);
     }
 
-
-    if (groupManager==NULL)
+    if (groupManager==nullptr)
     {
-        VERBOSE(sout << "Linking all contacts to Scene"<<sendl);
-        for (sofa::helper::vector<Contact::SPtr>::const_iterator it = notStaticContacts.begin(); it!=notStaticContacts.end(); ++it)
+        msg_info_when(d_doPrintInfoMessage.getValue())
+            << "Linking all contacts to Scene" ;
+
+        for (helper::vector<Contact::SPtr>::const_iterator it = notStaticContacts.begin(); it!=notStaticContacts.end(); ++it)
         {
             (*it)->createResponse(scene);
         }
     }
     else
     {
-        VERBOSE(sout << "Create Groups "<<groupManager->getName()<<sendl);
+        msg_info_when(d_doPrintInfoMessage.getValue())
+           << "Create Groups "<<groupManager->getName();
+
         groupManager->createGroups(scene, notStaticContacts);
     }
 }
@@ -234,8 +261,11 @@ helper::set< std::string > DefaultPipeline::getResponseList() const
 
 void DefaultPipeline::draw(const core::visual::VisualParams* )
 {
-    if (!bDraw.getValue()) return;
+    if (!d_doDebugDraw.getValue()) return;
     if (!narrowPhaseDetection) return;
+
+//TODO(dmarchal): remove this code or reactivate or do a proper #ifdef
+//TODO(dmarchal): it makes also no sense to keep a 'draw' attribute while nothing is displayed.
 #if 0
     glDisable(GL_LIGHTING);
     glLineWidth(2);